+++ src/pages/about/components/SpeechBubble.module.scss
@@ -1,0 +1,172 @@
+.speechBubble {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  position: relative;
+  margin-bottom: 2vh;
+
+  .bubbleOval {
+    background-color: var(--neutral-white, #fbf9f4);
+    border-radius: 100px;
+    border: 4px solid #1e1e1e;
+  }
+
+  .bubble {
+    position: relative;
+
+    .bubbleContainer:first-child {
+      position: absolute;
+      transform: rotate(2.489deg);
+      opacity: 0.8;
+    }
+
+    .bubbleContainer:nth-child(3) {
+      top: 0;
+      position: absolute;
+      transform: rotate(-2.489deg);
+      opacity: 0.8;
+    }
+
+    .bubbleContainer:nth-child(2) {
+      position: relative;
+      z-index: 2;
+    }
+
+    @media screen and (max-width: 720px) {
+      .bubbleContainer:first-child {
+        transform: rotate(3.6deg);
+      }
+
+      .bubbleContainer:nth-child(3) {
+        transform: rotate(-3.6deg);
+      }
+    }
+  }
+
+  .bubbleContainer {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    width: clamp(700px, 60vw, 900px);
+
+    @media screen and (max-width: 720px) {
+      width: clamp(360px, 90vw, 760px);
+    }
+
+    .textContainer {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: center;
+      width: 95%;
+      height: 162px;
+
+      @media screen and (max-width: 1000px) {
+        height: 120px;
+      }
+
+      @media screen and (max-width: 720px) {
+        height: 110px;
+      }
+
+      .primaryText {
+        color: var(--neutral-black, #1e1e1e);
+        text-align: center;
+        font-family: Snoopy;
+        font-size: var(--font-sizes-h2);
+        font-weight: 400;
+        letter-spacing: -1.056px;
+        text-transform: uppercase;
+      }
+      .secondaryText {
+        color: var(--neutral-black, #1e1e1e);
+        text-align: center;
+        font-family: Groundnut;
+        font-size: var(--font-sizes-base);
+        font-weight: 400;
+        line-height: 30px; /* 150% */
+        letter-spacing: -0.4px;
+        z-index: 10;
+      }
+    }
+
+    .bubbleTail {
+      align-self: flex-end;
+      width: 91.147px;
+      height: 66.54px;
+      margin-top: -30px;
+      margin-right: 40px;
+      // z-index: -1;
+
+      &::before {
+        content: "";
+        z-index: 4;
+        background-color: var(--neutral-white, #fbf9f4);
+        position: absolute;
+        width: 78.2px;
+        height: 32px;
+        border-radius: 100px;
+        border: none;
+        margin-top: -4px;
+        margin-left: -5px;
+
+        @media screen and (max-width: 1000px) {
+          width: 80px;
+          margin-left: -5px;
+        }
+      }
+
+      @media screen and (max-width: 720px) {
+        width: 64.1px;
+        height: 57.5px;
+        margin-right: 12px;
+
+        &::before {
+          width: 62.2px;
+          height: 27px;
+          margin-top: -5px;
+          margin-left: -13px;
+        }
+      }
+    }
+
+    .bubbleExcess {
+      width: 51.65px;
+      height: 37.706px;
+      align-self: flex-end;
+
+      @media screen and (max-width: 720px) {
+        width: 27px;
+        height: 24px;
+      }
+    }
+  }
+
+  /* Snoopy */
+  .imageContainer {
+    position: absolute;
+    top: 40%;
+    left: 12.5%;
+    z-index: 999;
+
+    width: clamp(50px, 12vw, 128.387px);
+    height: clamp(54px, 18vw, 148px);
+    aspect-ratio: 72/83;
+
+    img {
+      width: 100%;
+      height: 100%;
+      object-fit: contain;
+    }
+
+    @media screen and (max-width: 1000px) {
+      top: 40%;
+    }
+
+    @media screen and (max-width: 550px) {
+      top: 50%;
+    }
+  }
+}
