+++ src/pages/about/org-culture.module.scss
@@ -37,179 +37,6 @@
   }
 }
 
-.speechBubble {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  position: relative;
-  margin-bottom: 2vh;
-
-  .bubbleOval {
-    background-color: var(--neutral-white, #fbf9f4);
-    border-radius: 100px;
-    border: 4px solid #1e1e1e;
-  }
-
-  .bubble {
-    position: relative;
-
-    .bubbleContainer:first-child {
-      position: absolute;
-      transform: rotate(2.489deg);
-      opacity: 0.8;
-    }
-
-    .bubbleContainer:nth-child(3) {
-      top: 0;
-      position: absolute;
-      transform: rotate(-2.489deg);
-      opacity: 0.8;
-    }
-
-    .bubbleContainer:nth-child(2) {
-      position: relative;
-      z-index: 2;
-    }
-
-    @media screen and (max-width: 720px) {
-      .bubbleContainer:first-child {
-        transform: rotate(3.6deg);
-      }
-
-      .bubbleContainer:nth-child(3) {
-        transform: rotate(-3.6deg);
-      }
-    }
-  }
-
-  .bubbleContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: clamp(800px, 60vw, 900px);
-
-    @media screen and (max-width: 720px) {
-      width: clamp(360px, 90vw, 760px);
-    }
-
-    .textContainer {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      justify-content: center;
-      width: 95%;
-      height: 162px;
-
-      @media screen and (max-width: 1000px) {
-        height: 120px;
-      }
-
-      @media screen and (max-width: 720px) {
-        height: 110px;
-      }
-
-      .primaryText {
-        color: var(--neutral-black, #1e1e1e);
-        text-align: center;
-        font-family: Snoopy;
-        font-size: var(--font-sizes-h2);
-        font-weight: 400;
-        letter-spacing: -1.056px;
-        text-transform: uppercase;
-      }
-      .secondaryText {
-        color: var(--neutral-black, #1e1e1e);
-        text-align: center;
-        font-family: Groundnut;
-        font-size: var(--font-sizes-base);
-        font-weight: 400;
-        line-height: 30px; /* 150% */
-        letter-spacing: -0.4px;
-        z-index: 10;
-      }
-    }
-
-    .bubbleTail {
-      align-self: flex-end;
-      width: 91.147px;
-      height: 66.54px;
-      margin-top: -30px;
-      margin-right: 40px;
-      // z-index: -1;
-
-      &::before {
-        content: "";
-        z-index: 4;
-        background-color: var(--neutral-white, #fbf9f4);
-        position: absolute;
-        width: 78.2px;
-        height: 32px;
-        border-radius: 100px;
-        border: none;
-        margin-top: -4px;
-        margin-left: -5px;
-
-        @media screen and (max-width: 1000px) {
-          width: 80px;
-          margin-left: -5px;
-        }
-      }
-
-      @media screen and (max-width: 720px) {
-        width: 64.1px;
-        height: 57.5px;
-        margin-right: 12px;
-
-        &::before {
-          width: 62.2px;
-          height: 27px;
-          margin-top: -5px;
-          margin-left: -13px;
-        }
-      }
-    }
-
-    .bubbleExcess {
-      width: 51.65px;
-      height: 37.706px;
-      align-self: flex-end;
-
-      @media screen and (max-width: 720px) {
-        width: 27px;
-        height: 24px;
-      }
-    }
-  }
-
-  /* Snoopy */
-  .imageContainer {
-    position: absolute;
-    top: 40%;
-    left: 12.5%;
-    z-index: 999;
-
-    //     width: 50.691px;
-    // height: 54.271px;
-    width: clamp(50px, 12vw, 128.387px);
-    height: clamp(54px, 18vw, 148px);
-    aspect-ratio: 72/83;
-
-    img {
-      object-fit: contain;
-    }
-
-    @media screen and (max-width: 1000px) {
-      top: 40%;
-    }
-
-    @media screen and (max-width: 720px) {
-      top: 60%;
-    }
-  }
-}
-
 .hero {
   display: flex;
   align-items: center;
@@ -319,7 +146,7 @@
     .characters {
       display: flex;
       justify-content: center;
-      width: clamp(235px, 40vw, 530.556px);
+      width: clamp(235px, 50vw, 530.556px);
 
       .imageContainer {
         display: block;
@@ -370,7 +197,7 @@
     display: grid;
     justify-content: center;
     max-width: 1440px;
-    grid-template-columns: repeat(4, 20vw);
+    grid-template-columns: repeat(4, minmax(20vw, 292px));
     gap: 1vw;
 
     @media screen and (max-width: 1000px) {
@@ -388,7 +215,7 @@
       border: 2px solid var(--neutral-black);
 
       width: 100%;
-      height: 20vw;
+      height: minmax(20vw, 292px);
       padding: 10%;
 
       @media screen and (max-width: 1000px) {
@@ -448,10 +275,10 @@
 
   .imageContainer {
     border-bottom: 4px solid #1e1e1e;
-    background:
-      url("../../../public/noise.svg") lightgray 20.312px 0px / 89.752% 100% no-repeat,
-      url("../../../public/noise.svg") lightgray -198.826px -226.705px / 236.356% 228.337% no-repeat,
-      url("../../../public/noise.svg") lightgray 50% / cover no-repeat;
+    // background:
+    //   url("../../../public/noise.svg") lightgray 20.312px 0px / 89.752% 100% no-repeat,
+    //   url("../../../public/noise.svg") lightgray -198.826px -226.705px / 236.356% 228.337% no-repeat,
+    //   url("../../../public/noise.svg") lightgray 50% / cover no-repeat;
     background-blend-mode: overlay, normal, normal;
     border: 4px solid var(--neutral-black, #1e1e1e);
     box-shadow: -8px 8px 0 0 var(--neutral-black, #1e1e1e);
@@ -486,7 +313,7 @@
 
         margin-left: calc(-1 * var(--left-margin));
         margin-top: -24px;
-        padding: 2%;
+        padding: 5%;
 
         background-color: var(--neutral-white);
         border: 2.208px solid var(--neutral-black, #1e1e1e);
